refactor: A possible performance boost to the data retrieval from the storage
I decided to make a short break from Federation and make the refactoring I thought about for some time now.
I guess it should help with large conversations, but since it trades memory for speed, it's possible that
we will hit the GC limit and the performance boost will be wasted. Anyway, I'd like to try it out.

There are a few changes:
1. The simplest one. I added "final" to our storage classes declarations. From the work on signals I know
   that it actually helps the compiler a bit - the generated bytecode is a bit shorter. If we ever need
   to create a subclass of one of those classes, we can simply remove the "final" keyword from it.
   I also used this opportunity to split very long lines of those declarations into a few shorter ones.
2. In `UserStorage` I switched from a fixed size cache to an unlimited size cache. If the user has large
   conversations, it's possible that he or she will hit the cache limit for users which slows down
   database access.
3. The most complex change. Currently we have two ways to get all data in the storage as one big map/sequence:
   directly, through the `list()` method, or through a signal `contents` that keeps copy of data and gets updated
   with every update to the underlying storage. The signal ia there anyway so I decided to use it instead of
   the `list()` method, hoping for better performance. So, `list()` is gone, and instead we have `values` which
   returns a vector of data taken from `contents`, and `keySet`, which returns the set of keys (d'oh) from
   `contents` - in some cases this is all that we need and currently we retrieve all data with `list()` and
   then use only keys.
4. The old `list()` method returned a `Seq` which is a trait and its implementation is unsure. The new method
   `values` returns a `Vector` so we can be sure that iterating through data will be O(n).

You may ask, if there is a cache already in the storage, then shouldn't it be all the same which one we use?
And my answer is - I guess it should? But we have `contents` anyway and right now, after all the work with
signals, I'm inclined to believe that this signal is pretty fast. I made some rough tests on my device and
the dev build and the initial sync time for a reasonably big client got around 10% faster with those changes.
But of course that may be a coincidence. QA tests on large conversations will give us a better idea.