fix: increase CI emulator data partition size.
CI for the previous commit was failing due to an error regarding insufficient
disk space in the emulator when installing the application for running emulator
tests.

Some spelunking in the docker-android git repository (from which the emulator
containers are built) suggests that we can set a DATAPARTITION environment
variable on the docker command line, which will tell the containerised emulator
to create an appropriately sized data partition.