fix(notification): refactor  connections notifications to prevent ANR (AR-2898) (#1295)
* fix(notification): refactor  connections notifications to prevent ANR
To manage the notifications, we use the conversationId as the identifier. When a user go the connection request, we were passing the userId instead of passing the conversationId to the CurrentScreenManager then we had to fetch the conversationId by the userId from the DB in a insufficient way! In this fix we pass the conversation Id from userProfileScreen. The previous solution was causing ANR issues on the main thread [the main thread will be fixed in another PR]